+++ resources/js/components/RegistroHorasComponent.vue
@@ -366,17 +366,35 @@ export default {
         },
         eliminarSRDLetra(item, index)
         {
-            axios.delete(`/srd_letras/${item.id}`)
-                .then(()=>{
-                    this.$root.$emit('bv::hide::modal', 'modal-tabla', '#focusThisOnClose');
-                });
+            this.$bvModal.msgBoxConfirm("¿Quiere eliminar?",{
+                okVariant: 'danger',
+                okTitle: 'Eliminar',
+                cancelTitle: 'Cancelar'
+            }).then(value=> {
+                if( value === true )
+                {
+                    axios.delete(`/srd_letras/${item.id}`)
+                        .then(()=>{
+                            this.$root.$emit('bv::hide::modal', 'modal-tabla', '#focusThisOnClose');
+                        });
+                }
+            })  
         },
         eliminarSRDProyecto(item, index)
         {
-            axios.delete(`/srd_proyectos/${item.id}`)
-                .then(()=>{
-                    this.$root.$emit('bv::hide::modal', 'modal-tabla', '#focusThisOnClose');
-                });   
+            this.$bvModal.msgBoxConfirm("¿Quiere eliminar?",{
+                okVariant: 'danger',
+                okTitle: 'Eliminar',
+                cancelTitle: 'Cancelar'
+            }).then(value=> {
+                if( value === true )
+                {
+                    axios.delete(`/srd_proyectos/${item.id}`)
+                        .then(()=>{
+                            this.$root.$emit('bv::hide::modal', 'modal-tabla', '#focusThisOnClose');
+                        }); 
+                }
+            })   
         },
         añadirLetra(item)
         {
